requirements.txt
streamlit
torch
torchvision
Pillow
requests
streamlit-chat
google-generativeai
